Update and Add links to the symbolic opset files (#124)
Hi,

The original link to symbolic.py is no longer working since now there are multiple files for definitions of symbolic operator functions. Perhaps providing the links to the helper file and to the new opset files is the best fix? I also add the link to torch/csrc/autograd/generated/VariableType.h, for the sake of convenience.